Move general components to `ui`
Some components in `components[/util]` are general in nature, relating to web page presentation and behavior. The term "ui" captures this better so we move the "utility components" to such a module under the root.

The `DownloadInfoTooltip` component is more closely tied to the "business" so we move this to the `components` module.
