+++ WebRandomizer/ClientApp/src/components/NavMenu.jsx
@@ -1,10 +1,10 @@
 import React, { useState, useContext } from 'react';
 import { Link } from 'react-router-dom';
 import styled from 'styled-components';
-import classNames from 'classnames';
-
 import { Container, Navbar, NavbarBrand, NavbarToggler, Collapse, Nav, NavItem, NavLink } from 'reactstrap';
 import { UncontrolledDropdown, DropdownToggle, DropdownMenu, DropdownItem } from 'reactstrap';
+
+import classNames from 'classnames';
 
 import { GameTraitsCtx } from '../game/traits';
