+++ WebRandomizer/ClientApp/src/ui/BootstrapIcon.jsx
