+++ WebRandomizer/ClientApp/src/ui/PlainList.jsx
