+++ WebRandomizer/ClientApp/src/ui/PrefixInputGroup.jsx
@@ -1,7 +1,7 @@
 ﻿import React from 'react';
 import { InputGroup, InputGroupAddon, InputGroupText } from 'reactstrap';
 
-export default function PrefixedIndexGroup(props) {
+export default function PrefixInputGroup(props) {
     const { className, prefixClassName, prefix, children } = props;
     return <InputGroup className={className}>
         <InputGroupAddon className={prefixClassName} addonType="prepend">
